Move StorageService, StorageCredentials from client/ -> urbit/
